Use NVD3 to generate interactive SVG graphs and add a streamgraph.
The .dat files used for GNUPlot are replaced with JSON. The data is
preprocessed to minimize the amount of Javascript processing.

This requires fetch API to work so a somewhat modern browser with
javascript is needed.

Streamgraph shows the number of commits per month, so inactive authors
at a given time will not be visible. This gives a better overview of
when an author is active thant hte previous GNUplot stacked graph.
However, it does not give a view of the evolution over time of the
commit count. So, restore the commits per author graph from gitstats
doing that.